+++ lib/canary/plugs.ex
@@ -119,7 +119,7 @@ defmodule Canary.Plugs do
           fetch_resource(conn, opts)
       end
 
-    Plug.Conn.assign(conn, get_resource_name(conn, opts), loaded_resource)
+    Plug.Conn.assign(conn, get_resource_name(action, opts), loaded_resource)
   end
 
   @doc """
@@ -275,19 +275,12 @@ defmodule Canary.Plugs do
     action = get_action(conn)
     is_persisted = persisted?(opts)
 
-    non_id_actions =
-      if opts[:non_id_actions] do
-        Enum.concat([:index, :new, :create], opts[:non_id_actions])
-      else
-        [:index, :new, :create]
-      end
-
     resource =
       cond do
         is_persisted ->
           fetch_resource(conn, opts)
 
-        action in non_id_actions ->
+        action in non_id_actions(opts) ->
           opts[:model]
 
         true ->
@@ -371,7 +364,7 @@ defmodule Canary.Plugs do
     do: conn
 
   defp purge_resource_if_unauthorized(%{assigns: %{authorized: false}} = conn, opts),
-    do: Plug.Conn.assign(conn, get_resource_name(conn, opts), nil)
+    do: Plug.Conn.assign(conn, get_resource_name(get_action(conn), opts), nil)
 
   defp fetch_resource(conn, opts) do
     repo = Application.get_env(:canary, :repo)
@@ -380,7 +373,7 @@ defmodule Canary.Plugs do
 
     get_map_args = %{String.to_atom(field_name) => get_resource_id(conn, opts)}
 
-    case Map.fetch(conn.assigns, get_resource_name(conn, opts)) do
+    case Map.fetch(conn.assigns, get_resource_name(get_action(conn), opts)) do
       :error ->
         repo.get_by(opts[:model], get_map_args)
         |> preload_if_needed(repo, opts)
@@ -404,7 +397,7 @@ defmodule Canary.Plugs do
   defp fetch_all(conn, opts) do
     repo = Application.get_env(:canary, :repo)
 
-    resource_name = get_resource_name(conn, opts)
+    resource_name = get_resource_name(get_action(conn), opts)
 
     # check if a resource is already loaded at the key
     case Map.fetch(conn.assigns, resource_name) do
@@ -427,33 +420,6 @@ defmodule Canary.Plugs do
     end
   end
 
-  defp persisted?(opts) do
-    !!Keyword.get(opts, :persisted, false) || !!Keyword.get(opts, :required, false)
-  end
-
-  defp get_resource_name(conn, opts) do
-    case opts[:as] do
-      nil ->
-        opts[:model]
-        |> Module.split()
-        |> List.last()
-        |> Macro.underscore()
-        |> pluralize_if_needed(conn, opts)
-        |> String.to_atom()
-
-      as ->
-        as
-    end
-  end
-
-  defp pluralize_if_needed(name, conn, opts) do
-    if get_action(conn) in [:index] and not persisted?(opts) do
-      name <> "s"
-    else
-      name
-    end
-  end
-
   defp handle_unauthorized(%{assigns: %{authorized: true}} = conn, _opts),
     do: conn
 
@@ -467,17 +433,10 @@ defmodule Canary.Plugs do
   defp handle_not_found(conn, opts) do
     action = get_action(conn)
 
-    non_id_actions =
-      if opts[:non_id_actions] do
-        Enum.concat([:index, :new, :create], opts[:non_id_actions])
-      else
-        [:index, :new, :create]
-      end
-
     is_required = required?(opts)
-    resource_name = Map.get(conn.assigns, get_resource_name(conn, opts))
+    resource_name = Map.get(conn.assigns, get_resource_name(action, opts))
 
-    if is_nil(resource_name) and (is_required or action not in non_id_actions) do
+    if is_nil(resource_name) and (is_required or action not in non_id_actions(opts)) do
       apply_error_handler(conn, :not_found_handler, opts)
     else
       conn